+++ src/main/java/com/example/food/dto/DishesDto.java
@@ -3,20 +3,21 @@ package com.example.food.dto;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 import java.io.Serializable;
+import java.math.BigInteger;
 
 public class DishesDto implements Serializable {
-    private Long id;
+    private BigInteger id;
     private String description;
     @NotNull
-    private Long mass;
+    private BigInteger mass;
     @Pattern(regexp = "^\\w{2,45}$")
     private String name;
 
-    public Long getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
 
@@ -28,11 +29,11 @@ public class DishesDto implements Serializable {
         this.description = description;
     }
 
-    public Long getMass() {
+    public BigInteger getMass() {
         return mass;
     }
 
-    public void setMass(Long mass) {
+    public void setMass(BigInteger mass) {
         this.mass = mass;
     }
